some change

+++ icetea_wasm/src/safe_math.rs
@@ -9,15 +9,15 @@ macro_rules! require {
 }
 
 pub trait SafeMath {
-  fn mul(&self, b: u32) -> u32;
-  fn div(&self, b: u32) -> u32;
-  fn sub(&self, b: u32) -> u32;
-  fn add(&self, b: u32) -> u32;
-  fn modulo(&self, b: u32) -> u32;
+  fn mul(&self, b: u128) -> u128;
+  fn div(&self, b: u128) -> u128;
+  fn sub(&self, b: u128) -> u128;
+  fn add(&self, b: u128) -> u128;
+  fn modulo(&self, b: u128) -> u128;
 }
 
-impl SafeMath for u32 {
-  fn mul(&self, b: u32) -> u32 {
+impl SafeMath for u128 {
+  fn mul(&self, b: u128) -> u128 {
     if *self == 0 || b == 0  {
       return 0;
     }
@@ -26,25 +26,25 @@ impl SafeMath for u32 {
     return c;
   }
 
-  fn div(&self, b: u32) -> u32 {
+  fn div(&self, b: u128) -> u128 {
     require!(b > 0);
     let c = *self / b;
     return c;
   }
 
-  fn sub(&self, b: u32) -> u32 {
+  fn sub(&self, b: u128) -> u128 {
     require!(b <= *self);
     let c = *self - b;
     return c;
   }
 
-  fn add(&self, b: u32) -> u32 {
+  fn add(&self, b: u128) -> u128 {
     let c = *self + b;
     require!(c >= *self);
     return c;
   }
 
-  fn modulo(&self, b: u32) -> u32 {
+  fn modulo(&self, b: u128) -> u128 {
     require!(b > 0);
     return *self % b;
   }
